Build Working API
API is fully built and working, currently in progress of handling bugReporter() for the assignment of an optional User at it's creation. Currently, if no User is included it goes to 'No User Exists, statement/flash message.

Also, need to assign the users by referencing the unique identifier, so that if the project user changes their name, the assigned project user is updated also. 
Idea: Don't take the user name, but the unique identifier you create.

Also, 'if' statement to ensure random identifier is always unique.