Finished First Draft
API Is Fully working and completes all the functionality expected of the test.

Testing is implemented throughout and works with one exception.

If a new Bug Report is made, then immediately following the redirect you search the Bug title to see the full details, it crashes, quoting line 54

"details = request.form['bugDetails'] or None"
KeyError = bugDetails